testing sData

+++ AssetMetadata/API_Versions.cpp
@@ -12,6 +12,6 @@ namespace Pipeline {
 		};
 
 		Enum getEnum(const char* name) { return Enum(findStringArrayIndex(name, names, count)); }
-		std::string getName(const Enum& value) { return names[value]; }
+		std::string getName(const Enum& value) { return names[(uint8)value]; }
 	}
 }
